Tab lock aims at the chest; remote swings interpolate.
Lock-on now pitches at the target so a bow or staff shot is not a dirt ball. The other wanderer's swing clock ticks on the client between snapshots.

+++ crates/client/src/combat.rs
@@ -1097,7 +1097,7 @@ pub fn pose_weapons(
                 control.pred_loadout,
             )
         } else if let Some((_, pose)) = remotes.iter().find(|(e, _)| *e == parent.parent()) {
-            (pose.action, pose.action_ticks as f32, pose.loadout)
+            (pose.action, pose.action_ticks, pose.loadout)
         } else {
             continue;
         };
